+++ src/uart.c
@@ -1,12 +1,12 @@
 #include <uart.h>
 
-void uart1_init(uint32_t uart_baudrate)
+void uart1_init(uint16_t uart_baudrate)
 {
     SCON = 0x50; // Set Mode1: 8-Bit UART with Variable Baud Rate
     
     timer2_mode0_1T_init(); 
     enable_timer2_mode0_as_uart1_baud_rate(true);
-    
+
     timer2_mode0_start(timer2_mode0_baudrate_to_ticks(uart_baudrate));
 }
